Fixed line start and end points, and added a buffer for vertex spawns.
Line start points were generated under the assumption that ellipses were drawn from the top left of the shape, so removing the + Dimen.vertexDim/2 calculation fixed it. A buffer of 75 pixels was added so that no vertex can spawn within 75 pixels (from the center of the ellipse) of an edge.